+++ frontend/src/journal/new_journal/component/JournalNewContainer.tsx
@@ -1,34 +1,15 @@
-import { useMutation } from '@apollo/client'
-import { CREATE_JOURNAL } from '../graphql/mutation'
-import { useContext, useState } from 'react'
-import { AuthContext } from '../../../App'
+import { JournalNewHooks } from '../hooks/JournalNewHooks'
 import NewJournalPresentational from './JournalNewPresentational'
 
-
 export const NewJournalContainer = () => {
-  const { currentUser } = useContext(AuthContext)
-  const [createJournal] = useMutation(CREATE_JOURNAL)
-
-  const [title, setTitle] = useState('')
-  const [content, setContent] = useState('')
-
-  const handleCreateJournal = async () => {
-    try {
-      const res = await createJournal({
-        variables: {
-          title: title,
-          content: content,
-          userId: currentUser?.id
-        }
-      })
-      setTitle('')
-      setContent('')
-      console.log(res)
-    } catch (err) {
-      console.log(err)
-    }
-  }
-  
+  const {
+    currentUser,
+    title,
+    content,
+    handleCreateJournal,
+    setTitle,
+    setContent
+  } = JournalNewHooks()
 
   return (
     <NewJournalPresentational
